# MULE-15325: BindingContext look up is showing significant performance