--- a/core/src/main/java/org/mule/runtime/core/internal/el/DefaultBindingContextBuilder.java
+++ b/core/src/main/java/org/mule/runtime/core/internal/el/DefaultBindingContextBuilder.java
@@ -6,12 +6,16 @@
  */
 package org.mule.runtime.core.internal.el;
 
-import static java.util.Collections.unmodifiableList;
+import static java.util.Collections.emptyList;
+import static java.util.Collections.unmodifiableCollection;
 import static java.util.Collections.unmodifiableMap;
-import static java.util.Optional.empty;
+import static java.util.Optional.of;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toMap;
+import static org.mule.runtime.api.el.BindingContextUtils.ATTRIBUTES;
+import static org.mule.runtime.api.el.BindingContextUtils.NULL_BINDING_CONTEXT;
+import static org.mule.runtime.api.el.BindingContextUtils.PAYLOAD;
+import static org.mule.runtime.api.el.BindingContextUtils.VARS;
 
 import org.mule.runtime.api.el.Binding;
 import org.mule.runtime.api.el.BindingContext;
@@ -21,87 +25,194 @@ import org.mule.runtime.api.metadata.TypedValue;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.function.Supplier;
 
 
 public class DefaultBindingContextBuilder implements BindingContext.Builder {
 
-  private Map<String, Supplier<TypedValue>> bindings;
-  private List<ExpressionModule> modules;
+  private BindingContext parent;
+
+  private TypedValue payloadBinding;
+  private TypedValue attributesBinding;
+  private Supplier<TypedValue> varsBinding;
+
+  private final Map<String, Supplier<TypedValue>> bindings = new HashMap<>();
+  private Collection<ExpressionModule> modules = null;
 
   public DefaultBindingContextBuilder() {
-    this.bindings = new HashMap<>();
-    this.modules = new ArrayList<>();
+    this.parent = NULL_BINDING_CONTEXT;
+    // this.bindings = new HashMap<>();
   }
 
   public DefaultBindingContextBuilder(BindingContext bindingContext) {
-    this.bindings = bindingContext.identifiers().stream().collect(toMap(id -> id, id -> () -> bindingContext.lookup(id).get()));
-    this.modules = new ArrayList<>(bindingContext.modules());
+    this.parent = bindingContext;
+
+    payloadBinding = bindingContext.lookup(PAYLOAD).orElse(null);
+    attributesBinding = bindingContext.lookup(ATTRIBUTES).orElse(null);
+    varsBinding = () -> bindingContext.lookup(VARS).orElse(null);
   }
 
   @Override
   public BindingContext.Builder addBinding(String identifier, TypedValue value) {
-    bindings.put(identifier, () -> value);
+    switch (identifier) {
+      case PAYLOAD:
+        payloadBinding = value;
+        break;
+      case ATTRIBUTES:
+        attributesBinding = value;
+        break;
+      case VARS:
+        varsBinding = () -> value;
+        break;
+      default:
+        bindings.put(identifier, () -> value);
+        break;
+    }
     return this;
   }
 
   @Override
   public BindingContext.Builder addBinding(String identifier, Supplier<TypedValue> lazyValue) {
-    bindings.put(identifier, lazyValue);
+    switch (identifier) {
+      case PAYLOAD:
+        payloadBinding = lazyValue.get();
+        break;
+      case ATTRIBUTES:
+        attributesBinding = lazyValue.get();
+        break;
+      case VARS:
+        varsBinding = lazyValue;
+        break;
+      default:
+        bindings.put(identifier, lazyValue);
+        break;
+    }
     return this;
   }
 
   @Override
   public BindingContext.Builder addAll(BindingContext context) {
-    context.identifiers().forEach(id -> bindings.put(id, () -> context.lookup(id).get()));
-    modules.addAll(context.modules());
+    if (parent == null) {
+      parent = context;
+
+      payloadBinding = context.lookup(PAYLOAD).orElse(null);
+      attributesBinding = context.lookup(ATTRIBUTES).orElse(null);
+      varsBinding = () -> context.lookup(VARS).orElse(null);
+    } else {
+      // context.identifiers().forEach(id -> bindings.put(id, () -> context.lookup(id).get()));
+      context.bindings().forEach(binding -> {
+        addBinding(binding.identifier(), binding.value());
+      });
+      if (!context.modules().isEmpty()) {
+        modules = context.modules();
+      }
+    }
     return this;
   }
 
   @Override
   public BindingContext.Builder addModule(ExpressionModule expressionModule) {
-    this.modules.add(expressionModule);
+    if (modules == null) {
+      modules = new ArrayList<>();
+    }
+    modules.add(expressionModule);
     return this;
   }
 
   @Override
   public BindingContext build() {
-    return new BindingContextImplementation(bindings, modules);
+    return new BindingContextImplementation(parent, unmodifiableMap(bindings),
+                                            payloadBinding, attributesBinding, varsBinding,
+                                            modules != null ? unmodifiableCollection(modules) : emptyList());
   }
 
-  private static class BindingContextImplementation implements BindingContext {
+  public static class BindingContextImplementation implements BindingContext {
 
-    private Map<String, Supplier<TypedValue>> bindings;
-    private List<ExpressionModule> modules;
+    private final BindingContext parent;
 
-    private BindingContextImplementation(Map<String, Supplier<TypedValue>> bindings, List<ExpressionModule> modules) {
-      this.bindings = unmodifiableMap(new HashMap<>(bindings));
-      this.modules = unmodifiableList(new ArrayList<>(modules));
+    private final Optional<TypedValue> payloadBinding;
+    private final Optional<TypedValue> attributesBinding;
+    private final Supplier<TypedValue> varsBinding;
+
+    private final Map<String, Supplier<TypedValue>> bindings;
+    private final Collection<ExpressionModule> modules;
+
+    private BindingContextImplementation(BindingContext parent, Map<String, Supplier<TypedValue>> bindings,
+                                         TypedValue payloadBinding, TypedValue attributesBinding,
+                                         Supplier<TypedValue> varsBinding,
+                                         Collection<ExpressionModule> modules) {
+      this.parent = parent;
+      this.bindings = bindings;
+
+      this.payloadBinding = payloadBinding != null ? of(payloadBinding) : parent.lookup(PAYLOAD);
+      this.attributesBinding = attributesBinding != null ? of(attributesBinding) : parent.lookup(ATTRIBUTES);
+      this.varsBinding = varsBinding;
+
+      this.modules = modules;
     }
 
     @Override
     public Collection<Binding> bindings() {
-      return bindings.entrySet().stream()
-          .map(entry -> new Binding(entry.getKey(), entry.getValue() != null ? entry.getValue().get() : null)).collect(toList());
+      final List<Binding> bindingsList = bindings.entrySet().stream()
+          .map(entry -> new Binding(entry.getKey(), entry.getValue() != null ? entry.getValue().get() : null))
+          .collect(toList());
+
+      bindingsList.addAll(parent.bindings());
+
+      payloadBinding.ifPresent(pb -> bindingsList.add(new Binding(PAYLOAD, payloadBinding.get())));
+      attributesBinding.ifPresent(pb -> bindingsList.add(new Binding(ATTRIBUTES, attributesBinding.get())));
+      if (varsBinding != null && varsBinding.get() != null) {
+        bindingsList.add(new Binding(VARS, varsBinding.get()));
+      }
+
+      return bindingsList;
     }
 
     @Override
     public Collection<String> identifiers() {
-      return bindings.keySet();
+      final Set<String> identifiers = new HashSet<>(bindings.keySet());
+
+      identifiers.addAll(parent.identifiers());
+
+      payloadBinding.ifPresent(pb -> identifiers.add(PAYLOAD));
+      attributesBinding.ifPresent(pb -> identifiers.add(ATTRIBUTES));
+      if (varsBinding != null && varsBinding.get() != null) {
+        identifiers.add(VARS);
+      }
+
+      return identifiers;
     }
 
     @Override
     public Optional<TypedValue> lookup(String identifier) {
-      final Supplier<TypedValue> supplier = bindings.get(identifier);
-      return supplier != null ? ofNullable(supplier.get()) : empty();
+      switch (identifier) {
+        case PAYLOAD:
+          return payloadBinding;
+        case ATTRIBUTES:
+          return attributesBinding;
+        case VARS:
+          if (varsBinding == null) {
+            return parent.lookup(VARS);
+          } else {
+            return of(varsBinding.get());
+          }
+        default:
+          final Supplier<TypedValue> supplier = bindings.get(identifier);
+          return supplier != null ? ofNullable(supplier.get()) : parent.lookup(identifier);
+      }
     }
 
     @Override
     public Collection<ExpressionModule> modules() {
-      return modules;
+      List<ExpressionModule> mods = new ArrayList<>();
+      mods.addAll(parent.modules());
+      mods.addAll(modules);
+      return mods;
     }
   }
 }